+++ apps/ai/src/server/trpc/route/jobs/jobs.ts
@@ -37,6 +37,18 @@ import { z } from "zod";
 
 const SESSION_METADATA_NAME = "session.json";
 
+
+
+// 分布式训练框架
+export const Framework = z.union([
+  z.literal("tensorflow"),
+  z.literal("pytorch"),
+  z.literal("mindspore"),
+]);
+
+export type FrameworkType = z.infer<typeof Framework>;
+
+
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 const ImageSchema = z.object({
   name: z.string(),
@@ -61,6 +73,7 @@ const TrainJobInputSchema = z.object({
   algorithm: z.number().optional(),
   image: z.number().optional(),
   remoteImageUrl: z.string().optional(),
+  framework: Framework.optional(),
   isDatasetPrivate: z.boolean().optional(),
   dataset: z.number().optional(),
   isModelPrivate: z.boolean().optional(),
@@ -74,6 +87,7 @@ const TrainJobInputSchema = z.object({
   memory: z.number().optional(),
   maxTime: z.number(),
   command: z.string(),
+  gpuType: z.string().optional(),
 });
 
 export type TrainJobInput = z.infer<typeof TrainJobInputSchema>;
@@ -118,14 +132,31 @@ procedure
   })
   .mutation(
     async ({ input, ctx: { user } }) => {
-      const { clusterId, trainJobName, isAlgorithmPrivate, algorithm, image, remoteImageUrl,
+      const { clusterId, trainJobName, isAlgorithmPrivate, algorithm, image, framework, remoteImageUrl,
         isDatasetPrivate, dataset, isModelPrivate, model, mountPoints = [], account, partition,
-        coreCount, nodeCount, gpuCount, memory, maxTime, command } = input;
+        coreCount, nodeCount, gpuCount, memory, maxTime, command, gpuType } = input;
       const userId = user.identityId;
 
       const host = getClusterLoginNode(clusterId);
       if (!host) {
         throw clusterNotFound(clusterId);
+      }
+
+      const client = getAdapterClient(clusterId);
+
+      // 检查是否存在同名的作业
+      const existingJobName = await asyncClientCall(client.job, "getJobs", {
+        fields: ["job_id"],
+        filter: {
+          users: [userId], accounts: [],states: [],jobName:trainJobName,
+        },
+      }).then((resp) => resp.jobs);
+
+      if (existingJobName.length) {
+        throw new TRPCError({
+          code: "CONFLICT",
+          message: `trainJobName ${trainJobName} is already existed`,
+        });
       }
 
       const em = await forkEntityManager();
@@ -145,6 +176,7 @@ procedure
       return await sshConnect(host, userId, logger, async (ssh) => {
 
         const homeDir = await getUserHomedir(ssh, userId, logger);
+
 
         mountPoints.forEach((mountPoint) => {
           if (mountPoint && !isParentOrSameFolder(homeDir, mountPoint)) {
@@ -174,7 +206,6 @@ procedure
         const entryScript = command;
         await sftpWriteFile(sftp)(remoteEntryPath, entryScript);
 
-        const client = getAdapterClient(clusterId);
         const reply = await asyncClientCall(client.job, "submitJob", {
           userId,
           jobName: trainJobName,
@@ -195,6 +226,8 @@ procedure
           // 第五个参数为数据集版本地址
           // 第六个参数为模型版本地址
           // 第七个参数为多挂载点地址，以逗号分隔
+          // 第八个参数为gpuType, 表示训练时硬件卡的类型，由getClusterConfig接口获取
+          // 第九个参数告知适配器 该镜像对应的AI训练框架 如 tensorflow, pytorch 等
           extraOptions: [
             JobType.TRAIN,
             "",
@@ -215,6 +248,10 @@ procedure
                 : modelVersion.path
               : "",
             mountPoints.join(","),
+            gpuType || "",
+            // 如果是单机训练,则训练框架为空，表明为普通训练，华为的卡单机训练也要传框架
+            // 如果nodeCount不为1但同时选定镜像又没有框架标签，该接口会报错
+            (nodeCount === 1 && !gpuType?.startsWith("huawei.com")) ? "" : framework || "",
           ],
         }).catch((e) => {
           const ex = e as ServiceError;
@@ -230,8 +267,8 @@ procedure
           sessionId: trainJobName,
           submitTime: new Date().toISOString(),
           image: {
-            name: existImage!.name,
-            tag: existImage!.tag,
+            name: remoteImageUrl || existImage!.name,
+            tag: existImage?.tag || "latest",
           },
           jobType: JobType.TRAIN,
         };