+++ apps/mis-web/src/pages/accounts/[accountName]/info.tsx
@@ -25,6 +25,7 @@ import {
   checkQueryAccountNameIsAdmin } from "src/pageComponents/accounts/checkQueryAccountNameIsAdmin";
 import { getAccounts } from "src/pages/api/tenant/getAccounts";
 import { Head } from "src/utils/head";
+import { moneyNumberToString } from "src/utils/money";
 
 type Props = SSRProps<{
   accountName: string;
@@ -69,10 +70,10 @@ export const AccountInfoPage: NextPage<Props> = requireAuth(
           </Tag>
         </Descriptions.Item>
         <Descriptions.Item label={t("common.accountBalance")}>
-          {balance.toFixed(2)} {t("common.unit")}
+          {moneyNumberToString(balance)} {t("common.unit")}
         </Descriptions.Item>
         <Descriptions.Item label={t("common.blockThresholdAmount")}>
-          {blockThresholdAmount.toFixed(2)} {t("common.unit")}
+          {moneyNumberToString(blockThresholdAmount)} {t("common.unit")}
         </Descriptions.Item>
       </Descriptions>