+++ apps/portal-server/src/utils/proxy.ts
@@ -10,6 +10,7 @@
  * See the Mulan PSL v2 for more details.
  */
 
+import { ClusterConfigSchema } from "@scow/config/build/cluster";
 import { loggedExec, sftpWriteFile } from "@scow/lib-ssh";
 import { dirname } from "path";
 import { configClusters } from "src/config/clusters";
@@ -17,12 +18,12 @@ import { config } from "src/config/env";
 import { sshConnect } from "src/utils/ssh";
 import { Logger } from "ts-log";
 
-export const setupProxyGateway = async (logger: Logger) => {
+export const setupProxyGateway = async (logger: Logger, activatedClusters: Record<string, ClusterConfigSchema>) => {
 
   let portalBasePath = config.PORTAL_BASE_PATH;
   if (!portalBasePath.endsWith("/")) { portalBasePath += "/"; }
 
-  for (const id of Object.keys(configClusters)) {
+  for (const id of Object.keys(activatedClusters)) {
 
     const proxyGatewayConfig = configClusters[id].proxyGateway;