+++ libs/web/src/extensions/ExtensionPage.tsx
@@ -14,6 +14,8 @@ import { joinWithUrl } from "@scow/utils";
 import { useRouter } from "next/router";
 import React, { useEffect, useRef } from "react";
 import { Head } from "src/components/head";
+import { getExtensionRouteQuery } from "src/extensions/common";
+import { extensionEvents } from "src/extensions/events";
 import { ExtensionManifestWithUrl, UiExtensionStoreData } from "src/extensions/UiExtensionStore";
 import { UserInfo } from "src/layouts/base/types";
 import { useDarkMode } from "src/layouts/darkMode";
@@ -74,31 +76,42 @@ export const ExtensionPage: React.FC<Props> = ({
     return <NotFoundPageComponent />;
   }
 
+  const [title, setTitle] = React.useState(config?.name ?? "Extension");
+
   const darkMode = useDarkMode();
 
-  const query = new URLSearchParams(
-    Object.fromEntries(Object.entries(rest).filter(([_, val]) => typeof val === "string")) as Record<string, string>,
-  );
+  const extensionQuery = getExtensionRouteQuery(darkMode.dark, currentLanguageId, user?.token);
 
-  if (user) {
-    query.set("scowUserToken", user.token);
-  }
-
-  query.set("scowDark", darkMode.dark ? "true" : "false");
-
-  query.set("scowLangId", currentLanguageId);
+  const query = new URLSearchParams({
+    ...Object.fromEntries(Object.entries(rest).filter(([_, val]) => typeof val === "string")),
+    ...extensionQuery,
+  });
 
   const url = joinWithUrl(config.url, "extensions", ...pathParts)
     + "?" + query.toString();
 
   const ref = useRef<HTMLIFrameElement>(null);
 
-  // 监听来自iframe内部网页发送的信息，设置iframe的height
   useEffect(() => {
     const messageHandler = (e: MessageEvent<any>) => {
 
-      if (e.data.type === "scow.extensionPageHeightChanged" && ref.current) {
-        ref.current.style.height = e.data.payload.height + "px";
+      if (!ref.current) {
+        return;
+      }
+
+      const event = extensionEvents.safeParse(e.data);
+
+      if (!event.success) {
+        console.log("SCOW received an invalid event from extension page. event: %s", JSON.stringify(e.data));
+        return;
+      }
+
+      const data = event.data;
+
+      if (data.type === "scow.extensionPageHeightChanged") {
+        ref.current.style.height = data.payload.height + "px";
+      } else if (data.type === "scow.extensionPageTitleChanged") {
+        setTitle(data.payload.title);
       }
     };
     window.addEventListener("message", messageHandler, false);
@@ -110,7 +123,7 @@ export const ExtensionPage: React.FC<Props> = ({
 
   return (
     <>
-      <Head title={config?.name ?? "Extension"} />
+      <Head title={title} />
       <FrameContainer>
         <IFrame
           ref={ref}