+++ libs/web/src/layouts/base/common.tsx
@@ -28,6 +28,7 @@ export const EXTERNAL_URL_PREFIX = ["http://", "https://"];
 
 export function createMenuItems(
   routes: NavItemProps[],
+  pathname: string,
   parentClickable: boolean,
 ) {
 
@@ -53,7 +54,7 @@ export function createMenuItems(
             }
           }
           : undefined,
-        children: createMenuItems(route.children, parentClickable),
+        children: createMenuItems(route.children, pathname, parentClickable),
       } as ItemType;
     }
 
@@ -75,24 +76,33 @@ export function createMenuItems(
     } as ItemType;
   }
 
-  const items = routes.map((r) => createMenuItem(r));
+  const items = routes
+    .filter((x) => !x.hideIfNotActive || match(x, pathname))
+    .map((r) => createMenuItem(r));
 
   return items;
 }
 
-export function calcSelectedKeys(links: NavItemProps[], pathname: string) {
+export function calcActiveKeys(links: NavItemProps[], pathname: string): Set<string> {
 
-  return links.reduce((prev, curr) => {
-    if (arrayContainsElement(curr.children)) {
-      prev.push(...calcSelectedKeys(curr.children, pathname));
+  const selectedKeys = new Set<string>();
+
+  for (const link of links) {
+    if (arrayContainsElement(link.children)) {
+      const childrenSelectedKeys = calcActiveKeys(link.children, pathname);
+      for (const childKey of childrenSelectedKeys) {
+        selectedKeys.add(childKey);
+      }
     }
+
     if (
-      (curr.path === "/" && pathname === "/") ||
-        (curr.path !== "/" && curr.path !== "" && match(curr, pathname))
+      link.children?.some((x) => selectedKeys.has(x.path)) ||
+      (link.path === "/" && pathname === "/") ||
+        (link.path !== "/" && link.path !== "" && match(link, pathname))
     ) {
-      prev.push(curr.path);
+      selectedKeys.add(link.path);
     }
+  }
 
-    return prev;
-  }, [] as string[]);
+  return selectedKeys;
 }