+++ libs/web/src/layouts/base/header/components.tsx
@@ -10,7 +10,7 @@
  * See the Mulan PSL v2 for more details.
  */
 
-import { Tooltip } from "antd";
+import { Tooltip, Typography } from "antd";
 import NextLink from "next/link";
 import { useEffect, useRef, useState } from "react";
 import { antdBreakpoints } from "src/layouts/base/constants";
@@ -36,6 +36,14 @@ const Link = styled(NextLink)`
   align-items: center;
 `;
 
+const TypographyLink = styled(Typography.Link)`
+  font-size: 18px !important;
+  display: flex;
+  flex-wrap: wrap;
+  overflow: hidden;
+  align-items: center;
+`;
+
 export const TextSpan = styled.span`
 `;
 
@@ -49,9 +57,10 @@ interface JumpToAnotherLinkProps {
   href: string;
   text: string;
   hideText?: boolean;
+  crossSystem?: boolean;
 }
 
-export const JumpToAnotherLink: React.FC<JumpToAnotherLinkProps> = ({ href, icon, text, hideText }) => {
+export const JumpToAnotherLink: React.FC<JumpToAnotherLinkProps> = ({ href, icon, text, hideText, crossSystem }) => {
 
   const [isSmallScreen, setIsSmallScreen] = useState(false);
 
@@ -70,10 +79,9 @@ export const JumpToAnotherLink: React.FC<JumpToAnotherLinkProps> = ({ href, icon
     };
   }, []);
 
-  return (
-
-    <LinkItem>
-      <Link href={href} ref={linkRef}>
+  const content = () => {
+    return (
+      <>
         {(hideText || isSmallScreen) ? (
           <Tooltip title={text}>
             <IconContainer>
@@ -89,8 +97,23 @@ export const JumpToAnotherLink: React.FC<JumpToAnotherLinkProps> = ({ href, icon
               {text}
             </TextSpan>
           </>
-        )}
-      </Link>
+        )}</>
+    );
+  };
+
+  return (
+
+    <LinkItem>
+      {
+        crossSystem ? (
+          <TypographyLink href={href} ref={linkRef}>
+            {content()}
+          </TypographyLink>
+        ) : (
+          <Link href={href} ref={linkRef}>
+            {content()}
+          </Link>
+        ) }
     </LinkItem>
 
   );